Solve Day11

+++ src/day11/day11.kt
@@ -4,41 +4,41 @@ import inputTextOfDay
 import testTextOfDay
 import kotlin.math.floor
 
-fun part1(input: String): ULong {
+fun part1(input: String): Long {
     return monkeyBusiness(input, 20, ::reduceWorryLevels)
 }
-fun part2(input: String): ULong {
+fun part2(input: String): Long {
     return monkeyBusiness(input, 10000, ::manageWorryLevels)
 }
 
-fun monkeyBusiness(input: String, rounds: Int, managedWorries: (ULong, UInt) -> ULong): ULong {
+fun monkeyBusiness(input: String, rounds: Int, managedWorries: (Long, Long) -> Long): Long {
     val monkeys = parseMonkeys(input)
-    val monkeyItems = monkeys.map { monkey -> monkey.first().split(", ").map { it.toULong() }.toMutableList() }
-    val monkeyInspections = MutableList(8) { 0UL }
+    val monkeyItems = monkeys.map { monkey -> monkey.first().split(", ").map { it.toLong() }.toMutableList() }
+    val monkeyInspections = MutableList(8) { 0L }
 
-    val monkeyDivisor = monkeys.map{ monkey -> monkey[2].toUInt() }.reduce { a, b -> a * b }
+    val monkeyDivisor = monkeys.map{ monkey -> monkey[2].toLong() }.reduce { a, b -> a * b }
 
     repeat(rounds) {
         for (monkey in monkeys.indices) {
             monkeyItems[monkey].forEach {  worryLevel ->
-                monkeyInspections[monkey] = monkeyInspections[monkey] + 1UL
+                monkeyInspections[monkey] = monkeyInspections[monkey] + 1L
                 val (_, operation, tests, testTrue, testFalse) = monkeys[monkey]
                 val (operator, operand) = operation.split(" ")
 
                 val worryLevelDuringInspection = when (operator) {
                     "*" -> worryLevel * when (operand) {
                         "old" -> worryLevel
-                        else -> operand.toULong()
+                        else -> operand.toLong()
                     }
                     else -> worryLevel + when (operand) {
                         "old" -> worryLevel
-                        else -> operand.toULong()
+                        else -> operand.toLong()
                     }
                 }
 
                 val worryLevelAfterInspection = managedWorries(worryLevelDuringInspection, monkeyDivisor)
 
-                when (worryLevelAfterInspection % tests.toUInt() == 0UL) {
+                when (worryLevelAfterInspection % tests.toInt() == 0L) {
                     true -> {
                         monkeyItems[testTrue.toInt()].add(worryLevelAfterInspection)
                     }
@@ -53,10 +53,10 @@ fun monkeyBusiness(input: String, rounds: Int, managedWorries: (ULong, UInt) -> 
     return monkeyInspections.sortedDescending().take(2).let { (a,b) -> a * b }
 }
 
-private fun reduceWorryLevels(worryLevel: ULong, monkeyDivisor: UInt) =
-    floor(worryLevel.toDouble()  / 3.0).toULong()
+private fun reduceWorryLevels(worryLevel: Long, monkeyDivisor: Long) =
+    floor(worryLevel.toDouble()  / 3.0).toLong()
 
-private fun manageWorryLevels(worryLevel: ULong, monkeyDivisor: UInt) =
+private fun manageWorryLevels(worryLevel: Long, monkeyDivisor: Long) =
     worryLevel % monkeyDivisor
 
 private fun parseMonkeys(input: String) = input.lines().chunked(7) { monkey ->
@@ -81,12 +81,12 @@ fun main() {
 
     println(part1(test))
     println(part1(input))
-    check(part1(test) == 10605UL)
-    check(part1(input) == 57838UL)
+    check(part1(test) == 10605L)
+    check(part1(input) == 57838L)
 
     println(part2(test))
     println(part2(input))
 
-    check(part2(test) == 2713310158UL)
-    check(part2(input) == 15050382231UL)
+    check(part2(test) == 2713310158L)
+    check(part2(input) == 15050382231L)
 }
